market_edge_finder_experiment/requirements.txt
# Market Edge Finder Experiment - Production Requirements
# Python 3.11+ required

# Core ML Libraries
torch>=2.1.0
lightgbm>=4.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0  # For Parquet support
fastparquet>=2023.10.0

# OANDA API (Official v20 Library ONLY)
v20>=3.0.25.0

# Async HTTP Client
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Database
# sqlite3 is built-in

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.7.0

# Logging & Monitoring
structlog>=23.2.0
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Visualization (Optional - for evaluation)
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Performance Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# HTTP Requests
requests>=2.31.0

# Time Series Analysis
statsmodels>=0.14.0

# Optional GPU Support
# For CUDA: pip install torch>=2.1.0+cu118 -f https://download.pytorch.org/whl/cu118/torch_stable.html
# For Mac M1/M2: Metal backend auto-detected by PyTorch